% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{pretty_dt}
\alias{pretty_dt}
\title{Prettify a datatable by using Semantic UI}
\usage{
pretty_dt(data, ...)
}
\arguments{
\item{data}{Any data frame.}

\item{...}{Any other DT::datatable parameters.}

\item{class}{Table class that default to Semantic UI CSS.}
}
\value{
Interactive HTML table (DT::datatable).
}
\description{
Extends DT::datatable and automatically formats dates and numbers
while preserving sorting and order of columns.
}
\details{
More about Semantic UI \url{https://semantic-ui.com/}
can be found here.
}
